+++ src/rann.rs
@@ -16,7 +16,7 @@ impl Rann {
         Self { layers }
     }
 
-    pub fn forward(&mut self, input: &Vec<i64>) -> Vec<i64> {
+    pub fn forward(&mut self, input: &Vec<f32>) -> Vec<f32> {
         let mut output = input.clone();
         for layer in self.layers.iter_mut() {
             output = layer.forward(&output);
@@ -34,7 +34,7 @@ mod test {
     fn rann_dims() {
         let output_size = 10;
         let mut rann = Rann::new(&vec![5, 4, 3, 7, 9, output_size]);
-        let input = vec![-1, 3, 5, 3, -6];
+        let input = vec![-1., 3., 5., 3., -6.];
         let output = rann.forward(&input);
         println!("Output: {:?}", output);
         assert_eq!(output.len(), output_size);
